+++ TrendingGiphyBot/TrendingGiphyBotModel/JobConfigContainer.cs
@@ -3,20 +3,20 @@
     public class JobConfigContainer
     {
         public decimal ChannelId { get; set; }
-        public int? Interval { get; set; }
+        public short? Interval { get; set; }
         public string Time { get; set; }
         public string RandomSearchString { get; set; }
         public short? MinQuietHour { get; set; }
         public short? MaxQuietHour { get; set; }
         public string Prefix { get; set; }
         public JobConfigContainer() { }
-        public JobConfigContainer(decimal channelId, int? interval, string time)
+        public JobConfigContainer(decimal channelId, short? interval, string time)
         {
             ChannelId = channelId;
             Interval = interval;
             Time = time;
         }
-        public JobConfigContainer(JobConfigContainer basedOn, int? interval, string time)
+        public JobConfigContainer(JobConfigContainer basedOn, short? interval, string time)
         {
             ChannelId = basedOn.ChannelId;
             Interval = interval;
